Default to displaying results of just last build
This change modifies ninjatracing so that by default it just displays
the timing information from the last build. When multiple builds have
been done this is generally more useful. If the .ninja_log file only
includes data from one build then this makes no difference.

The --showall (-a) switch was added to allow the old behavior to be
requested.

This resolves issue #3.